requirements.txt
Pillow==10.0.0
Pillow==10.1.0
Requests==2.31.0
streamlit==1.28.1
streamlit_lottie==0.0.5
